+++ theme/src/main/java/de/baumann/thema/helpers/CustomListAdapter_Wallpaper.java
@@ -8,27 +8,28 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ArrayAdapter;
+import android.widget.ImageView;
 import android.widget.TextView;
 
 import de.baumann.thema.R;
 
-public class CustomListAdapter extends ArrayAdapter<String> {
+public class CustomListAdapter_Wallpaper extends ArrayAdapter<String> {
 
     private final Activity context;
     private final String[] itemTITLE;
-    private final String[] itemURL;
     private final String[] itemDES;
-    @SuppressWarnings("FieldCanBeLocal")
-    private final String[] itemFN;
+    @SuppressWarnings({"unused", "FieldCanBeLocal"})
+    private final int[] itemPATH;
+    private final int[] itemPATHPREV;
 
-    public CustomListAdapter(Activity context, String[] itemTITLE, String[] itemURL, String[] itemDES, String[] itemFN) {
+    public CustomListAdapter_Wallpaper(Activity context, String[] itemTITLE, String[] itemDES, int[] itemPATH, int[] itemPATHPREV) {
         super(context, R.layout.sound_item_list, itemTITLE);
 
         this.context=context;
         this.itemTITLE=itemTITLE;
-        this.itemURL=itemURL;
         this.itemDES=itemDES;
-        this.itemFN=itemFN;
+        this.itemPATH=itemPATH;
+        this.itemPATHPREV=itemPATHPREV;
     }
 
     @NonNull
@@ -37,16 +38,16 @@ public class CustomListAdapter extends ArrayAdapter<String> {
         if (rowView == null) {
             LayoutInflater infInflater = (LayoutInflater) this.context
                     .getSystemService(Context.LAYOUT_INFLATER_SERVICE);
-            rowView = infInflater.inflate(R.layout.sound_item_list, parent, false);
+            rowView = infInflater.inflate(R.layout.wallpaper_item_list, parent, false);
         }
 
         TextView textTITLE = (TextView) rowView.findViewById(R.id.textView_title);
-        TextView textURL = (TextView) rowView.findViewById(R.id.textView_url);
         TextView textDES = (TextView) rowView.findViewById(R.id.textView_des);
+        ImageView preview = (ImageView) rowView.findViewById(R.id.preview);
 
         textTITLE.setText(itemTITLE[position]);
-        textURL.setText(itemURL[position]);
         textDES.setText(itemDES[position]);
+        preview.setImageResource(itemPATHPREV[position]);
         return rowView;
     }
 }